refactor(cli): extract hooks command family into cli/commands/hooks.ts
